Create a custom hook to fetch the Games
